+++ src/components/layout/table/TableBody.tsx
@@ -6,11 +6,21 @@ import React from 'react'
 import { useTableDataContext } from './TableContext'
 import clsx from 'clsx'
 import { PropsUtil } from '@/src/utils/propsUtil'
+import { Visibility } from '../Visibility'
 
 export const TableBody = React.memo(function TableBodyVisual() {
-  const { table, onRowClick, isUsingFillerRows, fillerRow, pagination, rows } = useTableDataContext<unknown>()
-  const columns = table.getAllColumns()
-
+  const { table, onRowClick, onFillerRowClick, isUsingFillerRows, fillerRowCell, pagination, rows } = useTableDataContext<unknown>()
+  const pinnedColumnsLeft = table.getState().columnPinning?.left ?? []
+  const pinnedColumnsRight = table.getState().columnPinning?.right ?? []
+  let columnOrder = table.getState().columnOrder
+  columnOrder = [...pinnedColumnsLeft, ...columnOrder.filter(id => !pinnedColumnsLeft.includes(id) && !pinnedColumnsRight.includes(id)), ...pinnedColumnsRight]
+  const columnVisibility = table.getState().columnVisibility
+  const columns = columnOrder.map(id => {
+    const column = table.getColumn(id)
+    if (!column) return null
+    if (columnVisibility[id] === false) return null
+    return column
+  }).filter(Boolean)
   return (
     <tbody>
       {rows.map(row => {
@@ -34,19 +44,26 @@ export const TableBody = React.memo(function TableBodyVisual() {
           </tr>
         )
       })}
-      {isUsingFillerRows && range(pagination.pageSize - rows.length, { allowEmptyRange: true }).map((row, index) => {
-        return (
-          <tr key={'filler-row-' + index} className={clsx('table-body-filler-row')}>
-            {columns.map((column) => {
-              return (
-                <td key={column.id} className={clsx('table-body-filler-cell', column.columnDef.meta?.className)}>
-                  {fillerRow ? fillerRow(column.id, table) : (<FillerCell />)}
-                </td>
-              )
-            })}
-          </tr>
-        )
-      })}
+      <Visibility isVisible={isUsingFillerRows}>
+        {range(pagination.pageSize - rows.length, { allowEmptyRange: true }).map((index) => {
+          return (
+            <tr
+              key={'filler-row-' + index}
+              className={clsx('table-body-filler-row')}
+              onClick={() => onFillerRowClick?.(index, table)}
+              data-clickable={PropsUtil.dataAttributes.bool(!!onFillerRowClick)}
+            >
+              {columns.map((column) => {
+                return (
+                  <td key={column.id} className={clsx('table-body-filler-cell', column.columnDef.meta?.className)}>
+                    {fillerRowCell ? fillerRowCell(column.id, table) : (<FillerCell />)}
+                  </td>
+                )
+              })}
+            </tr>
+          )
+        })}
+      </Visibility>
     </tbody>
   )
 })